+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Use Kubernetes Events on `AWSCluster` CR for draining status updates.
+- Added the use the runtime/default seccomp profile.
 
 ### Changed