+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.1] - 2023-03-02
+
 ### Added
 
 - Use Kubernetes Events on `AWSCluster` CR for draining status updates.
@@ -86,7 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Push `node-operator` chart into `control-plane` catalog instead of quay.io.
 - Push `node-operator` app CRs into `<provider>-app-collection` repository.
 
-[Unreleased]: https://github.com/giantswarm/node-operator/compare/v1.4.2...HEAD
+[Unreleased]: https://github.com/giantswarm/node-operator/compare/v2.0.1...HEAD
+[2.0.1]: https://github.com/giantswarm/node-operator/compare/v1.4.2...v2.0.1
 [1.4.2]: https://github.com/giantswarm/node-operator/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/giantswarm/node-operator/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/giantswarm/node-operator/compare/v1.3.0...v1.4.0